Moving code from BatchProcessing and BatchUnionProcessing header files to impl files and cleaned up some includes
