+++ engine/CMakeLists.txt
@@ -255,6 +255,8 @@ set(SRC_FILES ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BlazingH
               ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/taskflow/graph.cpp
               ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BatchJoinProcessing.cpp
               ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BatchOrderByProcessing.cpp
+              ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BatchUnionProcessing.cpp
+              ${CMAKE_SOURCE_DIR}/src/execution_graph/logic_controllers/BatchProcessing.cpp
               ${CMAKE_SOURCE_DIR}/src/config/GPUManager.cu
               ${CMAKE_SOURCE_DIR}/src/operators/OrderBy.cpp
               ${CMAKE_SOURCE_DIR}/src/operators/GroupBy.cpp
