+++ engine/src/execution_graph/logic_controllers/PhysicalPlanGenerator.h
@@ -1,19 +1,18 @@
 #pragma once
 
-#include "BlazingColumn.h"
 #include "LogicPrimitives.h"
-#include "CacheMachine.h"
-#include "BatchProcessing.h"
 #include "BatchOrderByProcessing.h"
 #include "BatchAggregationProcessing.h"
 #include "BatchJoinProcessing.h"
 #include "BatchUnionProcessing.h"
-#include "io/DataLoader.h"
 #include "io/Schema.h"
 #include "utilities/CommonOperations.h"
+#include "parser/expression_utils.hpp"
+#include <boost/property_tree/ptree.hpp>
+#include <boost/property_tree/json_parser.hpp>
+#include <Util/StringUtil.h>
 
 using namespace fmt::literals;
-
 
 namespace ral {
 namespace batch {
